API v1 namespacing (#483)
* API v1 namespacing

This commit does NOT add new functionality. It merely makes a new path
/v1 for the v1 version of the API. The controllers fully inherit the
ones we have by default.

In the future, we should make a v2/ set of controllers, which can
coexist with the v1 version until it's stable enough to deprecate it.

* Move tests to v1 scope

* Rename old controllers to v1

* Move top-class comment

* Fix test rubocop
